# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CircleTrade Agent Stack API
  description: |
    The Circle Agent Stack API exposes CircleTrade AI's export intelligence pipeline as
    individually callable micro-agent endpoints. Each endpoint maps to a distinct intelligence
    module: DDTRS (compliance & HS classification), MeridianFlow (route optimisation),
    and TradeMatch (market & tariff intelligence).

    **Authentication**: All `/api/*` endpoints require an `X-API-Key` header.
    Demo key: `ct-demo-key-2026`
  version: 1.0.0
  contact:
    name: CircleTrade AI
    url: https://circletrade.ai
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: product_description is required
        timestamp:
          type: string
          format: date-time
...
    RouteResult:
      type: object
      properties:
        name: { type: string }
        score: { type: integer }
        transit_days: { type: integer }
        cost_index: { type: number }
        port_efficiency: { type: integer }
        political_risk: { type: string }
        regulatory_alignment: { type: string }
        primary_ports:
          type: array
          items: { type: string }
        recommendation: { type: string }
    RouteResponse:
      type: object
      properties:
        request_id: { type: string }
        query:
          type: object
          properties:
            origin: { type: string }
            destination: { type: string }
            product_category: { type: string, nullable: true }
        routes:
          type: array
          items:
            $ref: '#/components/schemas/RouteResult'
        best_route:
          $ref: '#/components/schemas/RouteResult'
        total_routes_found: { type: integer }
        source: { type: string }
        timestamp: { type: string, format: date-time }
    MarketMatch:
      type: object
      properties:
        market: { type: string }
        market_code: { type: string }
        region: { type: string }
        score: { type: integer }
        tariff: { type: number }
        import_growth: { type: number }
        gdp_growth: { type: number }
        ease_of_trade: { type: integer }
        category_fit: { type: string, enum: [HIGH, MODERATE] }
        recommendation: { type: string }
    MarketResponse:
      type: object
      properties:
        request_id: { type: string }
        product_category: { type: string }
        matches:
          type: array
          items:
            $ref: '#/components/schemas/MarketMatch'
        best_market:
          $ref: '#/components/schemas/MarketMatch'
        total_markets_evaluated: { type: integer }
        source: { type: string }
        timestamp: { type: string, format: date-time }
...
    PipelineStep:
      type: object
      properties:
        module: { type: string }
        output: { type: object }
    PipelineResponse:
      type: object
      properties:
        request_id: { type: string }
        customer: { type: string }
        origin_country: { type: string }
        category: { type: string }
        best_route:
          type: object
          properties:
            name: { type: string }
            score: { type: integer }
            transit_days: { type: integer }
            cost_index: { type: number }
        best_market:
          type: object
          properties:
            name: { type: string }
            code: { type: string }
            region: { type: string }
            score: { type: integer }
        tariff_rate: { type: number, nullable: true }
        risk_score: { type: integer }
        opportunity_score: { type: integer }
        license_required: { type: boolean }
        license_type: { type: string, nullable: true }
        sanctioned_destination: { type: boolean }
        recommendation: { type: string }
        next_action: { type: string }
        pipeline_steps:
          type: array
          items:
            $ref: '#/components/schemas/PipelineStep'
        timestamp: { type: string, format: date-time }
        source: { type: string }
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/route:
    post:
      summary: Optimise trade corridor / shipping route
      tags: [MeridianFlow]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [origin, destination]
              properties:
                origin:
                  type: string
                  example: UK
                destination:
                  type: string
                  example: UAE
                product_category:
                  type: string
                  example: textiles
      responses:
        '200':
          description: Route optimisation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/market:
    post:
      summary: Match product to target markets
      tags: [TradeMatch]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_category]
              properties:
                product_category:
                  type: string
                  example: textiles
                target_markets:
                  type: array
                  items: { type: string }
                  example: [UAE, SGP, GBR]
                  description: Optional filter — evaluates all markets if omitted
      responses:
        '200':
          description: Market match results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/pipeline:
    post:
      summary: Run the full CircleTrade intelligence pipeline
      description: |
        Executes all intelligence modules sequentially: HS code resolution,
        route optimisation, market matching, tariff lookup, compliance check,
        and opportunity scoring. Returns a unified result with a pipeline_steps trace.
      tags: [Pipeline]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customer_name, origin_country, products]
              properties:
                customer_name:
                  type: string
                  example: "Acme Exports Ltd"
                origin_country:
                  type: string
                  example: UK
                products:
                  type: array
                  items:
                    type: object
                    properties:
                      name: { type: string }
                      category: { type: string }
                      hs_code: { type: string }
                  example:
                    - name: "merino wool scarf"
                      category: textiles
                      hs_code: "6117.10"
                target_markets:
                  type: array
                  items: { type: string }
                  example: [UAE, SGP]
                budget:
                  type: number
                  example: 100000
                risk_tolerance:
                  type: string
                  enum: [low, medium, high]
                  example: medium
      responses:
        '200':
          description: Full pipeline result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...